feat: user moved to context one level higher
