+++ client/src/pages/Home/Home.test.tsx
@@ -3,7 +3,7 @@ import { render, screen } from '@testing-library/react';
 import Home from './Home';
 
 test('renders loading', () => {
-  render(<Home />);
+  render(<Home isLoading={true} isAuthenticated={true}/>);
   const loadingElement = screen.getByRole('presentation');
   expect(loadingElement).toBeInTheDocument();
 });
